requirements.txt
altair
numpy
pandas
pydeck
streamlit
Pillow
plotly
time
streamlit_js_eval